+++ packages/expo-router/src/modal/types.ts
@@ -1,9 +1,7 @@
-import type { NavigationProp, ParamListBase } from '@react-navigation/native';
 import { type ViewProps } from 'react-native';
 
 export interface ModalConfig {
   component: React.ReactNode;
-  parentNavigationProp: NavigationProp<ParamListBase>;
   uniqueId: string;
   animationType?: 'slide' | 'fade' | 'none';
   presentationStyle?: 'fullScreen' | 'overFullScreen' | 'pageSheet' | 'formSheet';